Changed extern call to assembly. Was previously declaredas int instead of double

+++ Assignment-2/pythagoras.c
@@ -35,7 +35,7 @@
 #include <stdio.h>
 #include <math.h>
 
-extern int triangle();
+extern double triangle();
 
 int main() {
     printf("%s", "Welcome to the Right Triangles program maintained by Aaron Cope.\n");
